pyproject.toml
[project]
name = "vfllwiki"
version = "0.1.0"
description = "Wiki data for VFLL Projects"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mkdocs>=1.6.1",
    "mkdocs-glightbox>=0.4.0",
    "mkdocs-material>=9.5.50",
    "pillow>=11.1.0",
]

[tool.uv.sources]
mkdocs-material = { git = "https://github.com/squidfunk/mkdocs-material" }
